fixup! [Wasm] eliminate getters of companion objects (in external interfaces) for inlined extension properties/methods (KT-60661, KT-76462)
