+++ compiler/ir/backend.wasm/src/org/jetbrains/kotlin/backend/wasm/WasmLoweringPhases.kt
@@ -264,6 +264,5 @@ val wasmLowerings: List<NamedCompilerPhase<WasmBackendContext, IrModuleFragment,
     ::WasmInlineObjectsWithPureInitializationLowering,
 
     ::WhenBranchOptimiserLowering,
-    ::WasmEliminateUnusedLocalVariables,
     ::IrValidationAfterLoweringPhase,
 )
